refact : LineSectionAcceptanceTest RestAssured 공통부분으로 변경
